Make display equations larger.
It turns out that the display equations (those that are not inline) were
displaying small because they were not getting wrapped in <p> tags.  By
changing the CSS to use the same font size in the body of the HTML
rather than just in paragraph tags, the display equations are now
rendering with the correct font size.